decomp progress: 3 functions to 100%, multiple partials
- WiiLight::UpdatePosition 4%->100% (3-case GX light setup w/ ASM_BLOCK)
- WiiEnumeration::Poll 2.27%->100% (state machine + loading flags)
- Archive::GetFileInfo 62.8%->100% (5 fixes: stack buf, line literal, decl order)
- Other incremental edits (ChordShapeGenerator, CharIKFingers, MultiMesh)

+++ src/system/char/CharIKFingers.cpp
@@ -279,7 +279,7 @@ void CharIKFingers::CalculateHandDest(int i1, int i2) {
                 if (curDesc.unk0) {
                     ::Add(curDesc.unk8, v194, v194);
                     Vector3 v1ac;
-                    Scale(v1a0, (float)(i - 2), v1ac);
+                    Scale(v1a0, i - 2.0f, v1ac);
                     ::Add(v1ac, v194, v194);
                     if (i == 0) {
                         Hmx::Matrix3 m158;
